Added Test Level 3 Ramps
Made a new level that makes player go back and forth using ramps and planes as walls so when you turn around the camera you can properly see.

Note, changed the default rotation of pickups so they are easier to move around in 3d space.

Note 2, added a cube object under the restart point to prevent clipping through the stage from resetting with a high velocity.